"equal" function for specifying exactly what value is expected
This is probably more esoteric than the other functions in this provider,
but can be used to compensate for overly-conservative results from
providers where there is some predictable rule for constructing the output
value based on information available during planning but nonetheless the
provider does not implement that rule.

This therefore allows module authors to implement such rules themselves, if
the risk tradeoff seems worth it to have fewer unknown values in the plan
and therefore more information to use when deciding whether the plan is
correct.

This could be particularly useful in environments where plans are subjected
to automated policy checks. A policy check could be set up to fail if a
crucial value is unknown, and then a module author could compensate for it
being unknown by using this function to record their assumption about what
the value would be and know that the plan will not apply successfully if
that assumption turns out to be wrong in practice.